🌸 update: seedAll and move route
