+++ routes/accountRouter.ts
@@ -1,12 +1,10 @@
 import express from 'express';
 
-import { getMe, getAllAccounts } from '@controllers/accountController';
-import { auth, adminAuth } from '@middleware/auth';
+import { getAllAccounts } from '@controllers/accountController';
+import { adminAuth } from '@middleware/auth';
 
 const accountRouter = express.Router();
 
 accountRouter.get('/list', adminAuth, getAllAccounts);
-
-accountRouter.get('/me', auth, getMe);
 
 export default accountRouter;
