+++ routes/indexRouter.ts
@@ -1,11 +1,14 @@
 import express from 'express';
 
-import { login, register } from '@controllers/accountController';
+import { getMe, login, register } from '@controllers/accountController';
+import { auth } from '@middleware/auth';
 
 const indexRouter = express.Router();
 
 indexRouter.get('/login', login);
 
 indexRouter.post('/register', register);
+
+indexRouter.get('/me', auth, getMe);
 
 export default indexRouter;
