+++ routes/seedRouter.ts
@@ -1,9 +1,13 @@
 import express from 'express';
 
-import { seedAccount } from '@controllers/accountController';
+import { seedAccount, seedIngredientType } from '@utils/seedUtil';
 
 const seedRouter = express.Router();
 
-seedRouter.get('/accounts', seedAccount);
+seedRouter.get('/', seedAccount(true), seedIngredientType());
+
+seedRouter.get('/accounts', seedAccount());
+
+seedRouter.get('/ingredients/types', seedIngredientType());
 
 export default seedRouter;
